improve EditorPanel with keyboard shortcuts
- make ItemEntry list navigatable by tab
- add F2 to rename functionality

+++ src/components/Panel/EditorPanel/EditorPanel.svelte
@@ -7,35 +7,56 @@
     import ItemEntry from './ItemEntry';
 
     const { flatItems, selectedItem } = kustomStore;
-    let hasFocus = false;
+    let hasFocus = false,
+        currentIndex = -1;
 
-    function onAddButtonClick() {
-        dialogStore.show(NewItemDialog);
+    const itemElList: { focus(): Promise<void> }[] = [];
+
+    function moveCurIndex(backward = false) {
+        currentIndex += backward ? -1 : 1;
+        if (currentIndex >= $flatItems.length) currentIndex = 0;
+        if (currentIndex < 0) currentIndex = $flatItems.length - 1;
+
+        selectedItem.set($flatItems[currentIndex]);
     }
 
-    function onTreeClick() {
-        hasFocus = true;
-    }
+    function onKeydown(e: KeyboardEvent) {
+        if (!hasFocus) return;
 
-    function onTreeOutsideClick() {
-        hasFocus = false;
-    }
-
-    function onKeyup(e: KeyboardEvent) {
-        if (hasFocus && $selectedItem && e.key === 'Delete') {
-            kustomService.removeItem($selectedItem);
+        if (e.key === 'Tab') {
+            e.preventDefault();
+            moveCurIndex(e.shiftKey);
+        } else if ($selectedItem) {
+            if (e.key === 'Delete') kustomService.removeItem($selectedItem);
+            else if (e.key === 'F2')
+                currentIndex >= 0 && itemElList[currentIndex].focus();
         }
     }
 </script>
 
-<svelte:window on:keyup={onKeyup} />
+<svelte:window on:keydown={onKeydown} />
 
-<div class="tree" on:click={onTreeClick} use:outsideClick={onTreeOutsideClick}>
-    {#each $flatItems as item}
-        <ItemEntry {item} />
+<div
+    class="tree"
+    on:click|stopPropagation={() => (hasFocus = true)}
+    use:outsideClick={() => {
+        hasFocus = false;
+        currentIndex = -1;
+    }}
+>
+    {#each $flatItems as item, i}
+        <ItemEntry
+            {item}
+            active={currentIndex === i}
+            on:click={() => {
+                currentIndex = i;
+                selectedItem.set($flatItems[i]);
+            }}
+            bind:itemEl={itemElList[i]}
+        />
     {/each}
     <div class="spacer" on:click={() => ($selectedItem = undefined)} />
-    <button class="btn is-icon" on:click={onAddButtonClick}
+    <button class="btn is-icon" on:click={() => dialogStore.show(NewItemDialog)}
         ><Icon icon="add" /></button
     >
 </div>
